requirements.txt
PyYAML>=6.0
tqdm>=4.60.0
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
